+++ rusoto/credential/src/static_provider.rs
@@ -9,19 +9,18 @@ use {AwsCredentials, CredentialsError, ProvideAwsCredentials};
 /// Provides AWS credentials from statically/programmatically provided strings.
 #[derive(Clone, Debug)]
 pub struct StaticProvider {
-    /// The AWS Access Key ID to use for authenticating to AWS.
-    aws_access_key_id: String,
-    /// The AWS Secret Access Key to use for authenticating to AWS.
-    aws_secret_access_key: String,
-    /// The optional token to use for authenticating to AWS.
-    token: Option<String>,
-    /// The time in seconds each issued token should be valid for.
+    /// AWS credentials
+    credentials: AwsCredentials,
+
+    /// The time in seconds for which each issued token is valid.
     valid_for: Option<i64>,
 }
 
 impl StaticProvider {
     /// Creates a new Static Provider. This should be used when you want to statically, or programmatically
     /// provide access to AWS.
+    ///
+    /// `valid_for` is the number of seconds for which issued tokens are valid.
     pub fn new(
         access_key: String,
         secret_access_key: String,
@@ -29,9 +28,12 @@ impl StaticProvider {
         valid_for: Option<i64>,
     ) -> StaticProvider {
         StaticProvider {
-            aws_access_key_id: access_key,
-            aws_secret_access_key: secret_access_key,
-            token: token,
+            credentials: AwsCredentials::new(
+                access_key,
+                secret_access_key,
+                token,
+                None
+            ),
             valid_for: valid_for,
         }
     }
@@ -39,31 +41,34 @@ impl StaticProvider {
     /// Creates a new minimal Static Provider. This will set the token as optional none.
     pub fn new_minimal(access_key: String, secret_access_key: String) -> StaticProvider {
         StaticProvider {
-            aws_access_key_id: access_key,
-            aws_secret_access_key: secret_access_key,
-            token: None,
+            credentials: AwsCredentials::new(
+                access_key,
+                secret_access_key,
+                None,
+                None
+            ),
             valid_for: None,
         }
     }
 
     /// Gets the AWS Access Key ID for this Static Provider.
     pub fn get_aws_access_key_id(&self) -> &str {
-        &self.aws_access_key_id
+        &self.credentials.key
     }
 
     /// Gets the AWS Secret Access Key for this Static Provider.
     pub fn get_aws_secret_access_key(&self) -> &str {
-        &self.aws_secret_access_key
+        &self.credentials.secret
     }
 
     /// Determines if this Static Provider was given a Token.
     pub fn has_token(&self) -> bool {
-        self.token.is_some()
+        self.credentials.token.is_some()
     }
 
     /// Gets The Token this Static Provider was given.
     pub fn get_token(&self) -> &Option<String> {
-        &self.token
+        &self.credentials.token
     }
 
     /// Returns the length in seconds this Static Provider will be valid for.
@@ -76,19 +81,19 @@ impl ProvideAwsCredentials for StaticProvider {
     type Future = FutureResult<AwsCredentials, CredentialsError>;
 
     fn credentials(&self) -> Self::Future {
-        ok(AwsCredentials::new(
-            self.aws_access_key_id.clone(),
-            self.aws_secret_access_key.clone(),
-            self.token.clone(),
-            self.valid_for.map(|v| Utc::now() + Duration::seconds(v)),
-        ))
+        let mut creds = self.credentials.clone();
+        creds.expires_at = self.valid_for.map(|v| Utc::now() + Duration::seconds(v));
+        ok(creds)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use futures::Future;
+    use std::thread;
+    use std::time;
 
+    use test_utils::{is_secret_hidden_behind_asterisks, SECRET};
     use ProvideAwsCredentials;
     use super::*;
 
@@ -124,8 +129,39 @@ mod tests {
         let finalized = result.unwrap();
         let expires_at = finalized.expires_at().unwrap().clone();
 
-        // Give a wide range of time, just incase there's somehow an immense amount of lag.
+        // Give a wide range of time, just in case there's somehow an immense amount of lag.
         assert!(start_time + Duration::minutes(100) < expires_at);
         assert!(expires_at < start_time + Duration::minutes(200));
+    }
+
+    #[test]
+    fn test_static_provider_expiration_time_is_recalculated() {
+        let provider = StaticProvider::new(
+            "fake-key".to_owned(),
+            "fake-secret".to_owned(),
+            None,
+            Some(10000),
+        );
+        let creds1 = provider.credentials().wait().unwrap();
+        thread::sleep(time::Duration::from_secs(1));
+        let creds2 = provider.credentials().wait().unwrap();
+        assert!(creds1.expires_at() < creds2.expires_at());
+    }
+
+    #[test]
+    quickcheck! {
+        fn test_static_provider_secrets_not_in_debug(
+            access_key: String,
+            token: Option<()>,
+            valid_for: Option<i64>
+        ) -> bool {
+            let provider = StaticProvider::new(
+                access_key,
+                SECRET.to_owned(),
+                token.map(|_| SECRET.to_owned()),
+                valid_for
+            );
+            is_secret_hidden_behind_asterisks(&provider)
+        }
     }
 }